requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
facebook-sdk==2.0.0
idna==2.6
mysqlclient==1.3.12
oauthlib==2.0.4
PyMySQL==0.7.11
pytz==2017.2
requests==2.18.4
urllib3==1.22
